LayoutTests/svg/animations/svgnumberoptionalnumber-animation-4-expected.txt

SVG 1.1 dynamic animation tests

Test 'by' animation of SVGNumberOptionalNumber.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS feGaussianBlur.stdDeviationX.animVal is 5
PASS feGaussianBlur.stdDeviationY.animVal is 5
PASS feGaussianBlur.stdDeviationX.baseVal is 5
PASS feGaussianBlur.stdDeviationY.baseVal is 5
PASS feGaussianBlur.stdDeviationX.animVal is 10
PASS feGaussianBlur.stdDeviationY.animVal is 10
PASS feGaussianBlur.stdDeviationX.baseVal is 5
PASS feGaussianBlur.stdDeviationY.baseVal is 5
PASS feGaussianBlur.stdDeviationX.animVal is 15
PASS feGaussianBlur.stdDeviationY.animVal is 15
PASS feGaussianBlur.stdDeviationX.baseVal is 5
PASS feGaussianBlur.stdDeviationY.baseVal is 5
PASS feGaussianBlur.stdDeviationX.animVal is 5
PASS feGaussianBlur.stdDeviationY.animVal is 5
PASS feGaussianBlur.stdDeviationX.baseVal is 5
PASS feGaussianBlur.stdDeviationY.baseVal is 5
PASS successfullyParsed is true

TEST COMPLETE
